requirements.txt
flask==2.0.1
werkzeug==2.0.3
replicate==0.20.0
python-dotenv==0.19.0
requests==2.26.0 